GITHUB-APD-186: Rename in FileStorage query instead of InMemory
