+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/back/Infrastructure/CommunicationChannel/LocalFilestorage/serenity-updates.json
