+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/back/Infrastructure/Framework/Symfony/Resources/config/announcement_list.yml
@@ -10,8 +10,8 @@ services:
     akeneo_communication_channel.handler.query.list_announcements:
         class: 'Akeneo\Platform\CommunicationChannel\Application\Announcement\Query\ListAnnouncementsHandler'
         arguments:
-            - '@akeneo_communication_channel.query.in_memory.find_announcement_items'
+            - '@akeneo_communication_channel.query.local_filestorage.find_announcement_items'
             - '@akeneo_communication_channel.query.dbal.find_viewed_announcement_ids'
 
-    akeneo_communication_channel.query.in_memory.find_announcement_items:
-        class: 'Akeneo\Platform\CommunicationChannel\Infrastructure\CommunicationChannel\InMemory\InMemoryFindAnnouncementItems'
+    akeneo_communication_channel.query.local_filestorage.find_announcement_items:
+        class: 'Akeneo\Platform\CommunicationChannel\Infrastructure\CommunicationChannel\LocalFilestorage\LocalFilestorageFindAnnouncementItems'
