+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/back/Infrastructure/Framework/Symfony/Resources/config/has_new_announcements.yml
@@ -9,8 +9,8 @@ services:
     akeneo_communication_channel.handler.query.has_new_announcements:
         class: 'Akeneo\Platform\CommunicationChannel\Application\Announcement\Query\HasNewAnnouncementsHandler'
         arguments:
-            - '@akeneo_communication_channel.query.in_memory.find_new_announcement_ids'
+            - '@akeneo_communication_channel.query.local_filestorage.find_new_announcement_ids'
             - '@akeneo_communication_channel.query.dbal.find_viewed_announcement_ids'
 
-    akeneo_communication_channel.query.in_memory.find_new_announcement_ids:
-        class: 'Akeneo\Platform\CommunicationChannel\Infrastructure\CommunicationChannel\InMemory\InMemoryFindNewAnnouncementIds'
+    akeneo_communication_channel.query.local_filestorage.find_new_announcement_ids:
+        class: 'Akeneo\Platform\CommunicationChannel\Infrastructure\CommunicationChannel\LocalFilestorage\LocalFilestorageFindNewAnnouncementIds'
